Add completedFutureFrom method
This method is to be used to create completed futures
but in a way that is safe from leaking exceptions.

Today, one can do something like this,

```java
CompletionStage<String> getValue() {
  return completedFuture(computeValue());
}
```

but if `computeValue()` would throw then this exception leaks
synchronously and will break.

To handle this, one can do

```java
CompletionStage<String> getValue() {
  try {
    return completedFuture(computeValue());
  } catch (Exception e) {
    return exceptionallyCompletedFuture(e);
  }
}
```

but it is hard and burdemsome to always remember to handle that
exception.

This patch proposes a new alternative, but using a checked supplier;

```java
CompletionStage<String> getValue() {
  // if an exception thrown then stage is exceptionally completed
  return completedFutureFrom(this::computeValue);
}
```