replace checkstyle.xml with spotify-checkstyle-config
In moving from Checkstyle 5.x to 8.x there are a lot of changes needed
to the config, and it seemed simpler to reuse the config in
https://github.com/spotify/spotify-checkstyle-config/ rather than update
the file in this repo for each change.

There appear to be no violations in the source code in this repo after
making this change, so it seems like the checkstyle config here was
already more or less in line with the main one.